+++ AsyncTest/LocalMaximaTest.cs
@@ -26,10 +26,10 @@ namespace AsyncTest
                 10, 10, 10,
                 10, 20
             };
-            var inList = new BlockingCollection<int>();
+            var inList = new BlockingCollection<uint>();
             foreach (var item in list)
             {
-                inList.Add(item);
+                inList.Add((uint)item);
             }
             inList.CompleteAdding();
             var outList = new BlockingCollection<int>();
@@ -45,10 +45,10 @@ namespace AsyncTest
         public void LocalMaximaTestCorrectness2()
         {
             var list = new List<int> { 1, 1, 10, 1, 1, 1, 1, 10, 10 };
-            var inList = new BlockingCollection<int>();
+            var inList = new BlockingCollection<uint>();
             foreach (var item in list)
             {
-                inList.Add(item);
+                inList.Add((uint)item);
             }
             inList.CompleteAdding();
             var outList = new BlockingCollection<int>();
